Improve positioning of "about" section blurbs
Previously, the about section element had a static fixed height,
excluding the height of the three icons and blurbs at the bottom, and
space was reserved for those through the use of a large static fixed top
margin on the following element.  This caused significant problems in
cases where the height of the reflowed blurb text exceeded the expected
height, and can also look bad when the text is shorter than expected,
resulting in a large expanse of white appearing beneath the blurbs.
This commit modifies the markup to make those blurb heights be
considered part of the about section, while maintaining the exact
positioning of most elements and backgrounds.  Several pixel constants
are still used in order to keep the site looking precisely as it did
before; in the future, it may be a good idea to make the sizing of
various other elements more dynamic as well.  The positioning of the
icons, blurbs, BRL-CAD logo, and white background behind the blurbs is,
to the pixel, the same as it was before, although the lower bound of the
white background behind the blurbs now resizes based on the calculated
height of the blurb, providing the desired functionality.  There is a
single deviation in appearance from the previous version: the thin black
bar between the aforementioned white background and the background of
the BRL-CAD logo is now a fixed height.  This compromise was accepted
due to the impossibility of reproducing the previous arrangement without
accepting influences on the section from markup in other regions:
previously, the bar was generated by the lower padding of the items in
the menu, and as such depended on the font size, baseline alignment,
etc. of those menu items.